Core changes to support a pending types package
