+++ Source/Common/EnumHelper.cs
