+++ Source/Common/PropertyValueSpecification.Partial.cs
