+++ Source/Common/SchemaSerializer.cs
