+++ Source/Common/TypeExtensions.cs
