+++ Source/Common/Values{T1,T2,T3,T4,T5,T6,T7}.cs
