+++ Source/Schema.NET.Pending/Schema.NET.Pending.csproj
@@ -8,8 +8,8 @@
   </PropertyGroup>
 
   <PropertyGroup Label="Package">
-    <Product>Schema.NET</Product>
-    <Description>Schema.org objects turned into strongly typed C# POCO classes for use in .NET. All classes can be serialized into JSON/JSON-LD and XML, typically used to represent structured data in the head section of html page.</Description>
+    <Product>Schema.NET.Pending</Product>
+    <Description>Schema.org objects (including pending types) turned into strongly typed C# POCO classes for use in .NET. All classes can be serialized into JSON/JSON-LD and XML, typically used to represent structured data in the head section of html page.</Description>
     <PackageTags>Schema;.NET;Schema.org;Schema.NET;Structured Data;Google Structured Data</PackageTags>
   </PropertyGroup>
